Create Practice-11.html
It's the 11th practice, where i'm doing practice on adding Favicon to the page's title. I'm also trying to play with the CSS Style attribute. And i have checked the Header, Fooder and the Nav tag. My NAV tag is so for not complete and i would like to play more with it...